chore: add staged-lint
