+++ src/components/SpyButton.ts
@@ -1,0 +1,15 @@
+import { Color } from "../theme";
+
+export enum SpyButtonSize {
+  SM = "sm",
+  MD = "md",
+  LG = "lg",
+  XL = "xl",
+}
+
+export type SpyButtonProp = {
+  label: string;
+  onClick: () => void;
+  color: Color;
+  size: SpyButtonSize;
+};
