+++ src/components/SpyButton.tsx
@@ -1,19 +1,6 @@
 import classNames from "classnames";
+import { SpyButtonProp, SpyButtonSize } from "./SpyButton";
 import { Color } from "../theme";
-
-export enum SpyButtonSize {
-  SM = "sm",
-  MD = "md",
-  LG = "lg",
-  XL = "xl",
-}
-
-export type SpyButtonProp = {
-  label: string;
-  onClick: () => void;
-  color: Color;
-  size: SpyButtonSize;
-};
 
 export default function SpyButton({
   label,
